LayoutTests/platform/wpe/storage/domstorage/events/storage-event-not-in-originator-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
Tests that the storage event gets fired in other frames but not the originating one

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sessionStorage.getItem(sessionStorageKey) is null
PASS sessionStorageKey === localStorageKey is false
PASS localStorage.getItem(localStorageKey) is null
FAIL Unexpected error: QuotaExceededError: The quota has been exceeded.
FAIL successfullyParsed should be true. Was false.
Some tests failed.

TEST COMPLETE
